linux-raspberrypi: Integrate Sulka metadata
Integrate the Sulka kernel hardening metadata to the linux-raspberrypi
recipe. Use the new sulka-kernel-hardening bbclass to integrate the
metadata. This allows using the kernel hardening metadata from Sulka
with minimal configuration required.

Add Raspberry Pi 4 specific metadata to override some the settings in
the defconfigs and kernel features that disable security features. This
allows achieving better hardening of the system, and demonstrates how
the Sulka kernel metadata can be modified if so required.